requirements.txt
torch==2.1.0
diffusers[torch]==0.24.0
transformers
Pillow
runpod
